moved functions in a new file
